requirements.txt
streamlit==0.75.0
opencv-python==4.5.1.48
matplotlib==3.3.3
pillow==8.1.0
numpy==1.19.5
pandas==1.2.1